Resolved ECR edge case issue, also some improvements to ECR
Realized that some of the fetched data from IMAPClient server will not be bound to the search terms `RFC822`, checked the dictionary key is there before attempting to parse the email. Also made marking of unread emails also occur in ECR exception handler just in case. Can handle changes in inbox quite well. Only small issue is emails marked as unread that are older than other unread emails will be ignored but for the use case of Wail Mail this is fine since you want real-time notifications as the emails arrive.